Refactor snp_list_generator.py
