+++ collapsevariants/collapse_variants.py
@@ -15,8 +15,8 @@ from general_utilities.mrc_logger import MRCLogger
 
 from utilities.collapse_logger import CollapseLOGGER
 from utilities.ingest_data import IngestData
-from utilities.parallelization_wrappers import generate_generic_masks, generate_snp_or_gene_masks, \
-    generate_genotype_matrices, update_log_file
+from utilities.parallelization_wrappers import generate_genotype_matrices, update_log_file
+from tool_parsers.mask_generators import generate_generic_masks, generate_snp_or_gene_masks
 from collapsevariants.snp_list_generator.snp_list_generator import SNPListGenerator
 
 # Set up the system logger – this is NOT the same as LOG_FILE below that records info about the filtering itself
